Break down MCPClient into a MCPClientExtension.
